+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 <!-- Keep them in alphabetical order -->
 ## Enhancements
+- [KOGITO-7840](https://issues.redhat.com/browse/KOGITO-7840) - Add validation on Workflow Metadata with admission webhooks
 
 ## Bug Fixes